ROOT patches to make it work with pga at reasonable performance
These may cause thread safety issues if used in a different context, but appear to work fine with the way pga uses ROOT. Some mutex locks that have been recently introduced in ROOT have been removed. Without these patches, the pga performance with ROOT 6.18 is unacceptable as compared to ROOT 6.10.

The minimizer is fixed to Minuit2 without using the ROOT plugin system, which also caused major performance issues with pga's multithreading. A separate static library compilation of Minuit2 is required for linking.

One assert in Minuit2 has been commented out because this occationally causes crashes with the energy drift fine tune fit. Perhaps this gets triggered in the case of a fitting issue. Because the fit is tried many times with different intitial parameters, this is probably not an issue for the data analysis. The cause of the problem is not yet known.